fixed a few bugs
